feat: macOS and Windows support
